@@ -1,14 +1,20 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import {createRoot} from 'react-dom/client';
+import 'normalize.css';
+import './index.css';
 import App from "./App";
 
 import {Provider} from "react-redux";
-import store from "./app/store";
+import { store } from './store';
 
-ReactDOM.render(
-  <Provider store={store}>
-    <App />
-  </Provider>, 
-  document.getElementById("root"));
+
+const container = document.getElementById('root');
+const root = createRoot(container);
+
+root.render(
+    <Provider store={store}>
+        <App tab='home' />
+    </Provider>
+);
 
 
